Stupid timing hack for Win7
The timing hack

There is a bug on Windows7 which makes CreateProcessW sometimes return corrupted handles.

I traced it down to the DuplicateTokenEx function call, specifically when there is not enough time between the supposedly synchronized/blocking DuplicateTokenEx call and the following CreateProcessW, CPW will fail without an error but returns corrupted handles in the PROCESS_INFORMATION structure.

This is the route those functions take in the user space:  
advapi32:DuplicateTokenEx -> kernelbase:DuplicateTokenEx -> ntdll:NtDuplicateToken -> kernel syscall  
kernel32:CreateProcessW -> kernel32:CreateProcessInternalW -> ntdll:NtCreateUserProcess -> kernel syscall  

There seems to be nothing wrong in the user space part, so something goes wrong inside the kernel syscalls when you are "too fast".

Due to the fact that most of the internal kernel stuff is obscure and undocumented by design (security through obscurity) and also my kernel debugging expertise being pretty much 404, this is where the story ends for now. It also doesn't help that since the bug is timing-related, I cannot get it to fail under a debugger which inherently slows everything down enough to make the bug disappear.

I couldn't replicate this bug happening on Windows 10 / 11. Maybe the bug was indeed fixed at some point, or maybe it's due to the later Windows iterations being just slower in general, the underlying race condition or whatever never happens.

Anyway, this is the band-aid for now. Intentionally adding ~70ms of extra delay works. This is not applied on Windows versions 10 or above. They are slow enough as-is.